Fix formatting long as %d in util/log.cpp
Changed log level type from long to int, since value is
expected to be a small integer in range [0-3].

Since codebase requires C++17 also replaced strtol with
std::from_chars(), which will parse string as int, and skip updating
the defaut value if VULKAN_WSI_DEBUG_LEVEL env var does not represent
int.

+++ util/log.hpp
@@ -45,7 +45,7 @@ namespace util
  * @param[in] format    A C-style formatting string.
  */
 
-void wsi_log_message(long level, const char *file, int line, const char *format, ...)
+void wsi_log_message(int level, const char *file, int line, const char *format, ...)
 #ifdef __GNUC__
    __attribute__((format(printf, 4, 5)))
 #endif
